// client/src/app/user.ts
export class User{
		_id?: string;
		user_id: string;
		admin: boolean;
		mobile: string;
		name: string;
		gender: string;
		email: string;
		login_password: string;
		currency: string;
		walletAmount: string;
		mobile_verified: string;
		email_verified: string;
		deleted: boolean;
		createdBy: string;
		createdAt: string;
		changedBy: string;
		changedAt: string;
}